v111c_Build_005 Enable MapEditor, AssetEditor, Dynamic redirection
This build adds better support and fully enables the Map Editor to be
used with more trees, it also is a first pass at addressing Asset Editor
issues while this mod is enabled (it will try to unload itself after
entering AE).  Totally reworked the detouring so that we can revert the
detoured methods back when needed.  Changed things so that when in
doubt, even while enabled upon unload\reload of a map\game we'll patch
everything back to normal, reset the TreeMaster Array's and then redo it
for the next load. Added extended verbose logging(off by default) and
option panel setting for it and persistent storage of that setting.